Update to Swift 4.1.

+++ Sources/StringParser.swift
@@ -70,7 +70,7 @@ public func string (_ s: String) -> Parser<Character, String> {
         guard input.startIndex < input.endIndex else {
             throw ParseError.Mismatch(input, s, "EOF")
         }
-        guard let endIndex = input.index(input.startIndex, offsetBy:Int64(count), limitedBy: input.endIndex) else {
+        guard let endIndex = input.index(input.startIndex, offsetBy: count, limitedBy: input.endIndex) else {
             throw ParseError.Mismatch(input, s, String(input))
         }
         let next = input[input.startIndex..<endIndex]
@@ -105,8 +105,8 @@ public func noneOf(_ strings: [String]) -> Parser<Character, Character> {
         }
         for string in strings {
             guard string.first == next else { continue }
-            let offset = Int64(string.count)
-            guard Int64(input.count) >= offset else { continue }
+            let offset = string.count
+            guard input.count >= offset else { continue }
             let endIndex = input.index(input.startIndex, offsetBy: offset)
             guard endIndex <= input.endIndex else { continue }
             let peek = input[input.startIndex..<endIndex]
@@ -136,7 +136,7 @@ public func parse <A> (_ p: Parser<Character, A>, _ s: String) throws -> A {
 
 public func print(error: ParseError<Character>, in s: String) {
     if case ParseError<Character>.Mismatch(let remainder, let expected, let actual) = error {
-        let index = s.index(s.endIndex, offsetBy: -Int(remainder.count))
+        let index = s.index(s.endIndex, offsetBy: -remainder.count)
         let (lineRange, row, pos) = position(of: index, in: s)
         let line = s[lineRange.lowerBound..<lineRange.upperBound].trimmingCharacters(in: CharacterSet.newlines)
